refactor : config 위치 변경
