+++ src/main/java/kr/co/pinup/config/WebClientConfig.java
@@ -1,4 +1,4 @@
-package kr.co.pinup.members.config;
+package kr.co.pinup.config;
 
 import io.netty.channel.ChannelOption;
 import org.springframework.context.annotation.Bean;
